rename(#9): 도메인별 패키지 분리
