+++ service/coupon-domain/src/main/java/com/couponify/coupondomain/domain/coupon/Quantity.java
@@ -1,4 +1,4 @@
-package com.couponify.coupondomain.domain;
+package com.couponify.coupondomain.domain.coupon;
 
 import jakarta.persistence.Embeddable;
 import lombok.Getter;
