+++ service/coupon-domain/src/main/java/com/couponify/coupondomain/domain/coupon/repository/CouponRepository.java
@@ -1,0 +1,10 @@
+package com.couponify.coupondomain.domain.coupon.repository;
+
+
+import com.couponify.coupondomain.domain.coupon.Coupon;
+
+public interface CouponRepository {
+
+  public Coupon save(Coupon coupon);
+
+}
