--- service/coupon-domain/src/main/java/com/couponify/coupondomain/domain/repository/CouponRepository.java
@@ -1,9 +1,0 @@
-package com.couponify.coupondomain.domain.repository;
-
-import com.couponify.coupondomain.domain.Coupon;
-
-public interface CouponRepository {
-
-  public Coupon save(Coupon coupon);
-
-}
