+++ service/coupon-domain/src/main/java/com/couponify/coupondomain/infrastructure/jpa/coupon/CouponEntity.java
@@ -1,4 +1,4 @@
-package com.couponify.coupondomain.infrastructure.jpa;
+package com.couponify.coupondomain.infrastructure.jpa.coupon;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -18,6 +18,7 @@ public class CouponEntity {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
+  @Column(name = "coupon_id")
   private Long id;
 
   @Column(length = 100, nullable = false)
